Index.jsp

+++ database-backup/DBFOODWEB.sql
@@ -11,23 +11,23 @@ USE DBFOODWEB
 GO
 
 CREATE TABLE City (
-    id              SMALLINT        PRIMARY KEY IDENTITY,
+    id              INT PRIMARY KEY IDENTITY,
     [name]          NVARCHAR(100)
 );
 GO
 
 CREATE TABLE District (
-    id              SMALLINT        PRIMARY KEY IDENTITY,
+    id              INT PRIMARY KEY IDENTITY,
     [name]          NVARCHAR(100),
-    city_id         SMALLINT        NOT NULL,
+    city_id         INT NOT NULL,
     FOREIGN KEY (city_id) REFERENCES City(id)
 );
 GO
 
 CREATE TABLE Ward (
-    id              SMALLINT        PRIMARY KEY IDENTITY,
+    id              INT PRIMARY KEY IDENTITY,
     [name]          NVARCHAR(100),
-    district_id     SMALLINT        NOT NULL,
+    district_id     INT NOT NULL,
     FOREIGN KEY (district_id) REFERENCES District(id)
 );
 GO
@@ -39,9 +39,9 @@ CREATE TABLE Accounts (
     email			NVARCHAR(255) NOT NULL,
     phone           NVARCHAR(20),
     [address]       NVARCHAR(255),
-    ward_id         SMALLINT,
-    district_id     SMALLINT,
-    city_id         SMALLINT,
+    ward_id         INT,
+    district_id     INT,
+    city_id         INT,
     [role]          NVARCHAR(20) DEFAULT 'user',
     [status]        BIT NOT NULL DEFAULT 1,
     FOREIGN KEY (city_id) REFERENCES City(id),
@@ -51,8 +51,8 @@ CREATE TABLE Accounts (
 GO
 
 CREATE TABLE Ingredients (
-    id              SMALLINT        PRIMARY KEY IDENTITY,
-    [name]          NVARCHAR(255)   NOT NULL,
+    id              INT PRIMARY KEY IDENTITY,
+    [name]          NVARCHAR(255) NOT NULL,
     unit            NVARCHAR(50),
     price           MONEY,
     total_quantity  INT,
@@ -74,7 +74,7 @@ CREATE TABLE Menu (
 GO
 
 CREATE TABLE Dishes (
-    id              SMALLINT        PRIMARY KEY IDENTITY,
+    id              INT PRIMARY KEY IDENTITY,
     [name]          NVARCHAR(255),
     price           MONEY,
     [status]        BIT,
@@ -85,18 +85,19 @@ GO
 
 CREATE TABLE Menu_Dishes (
     menu_id         INT NOT NULL,
-    dish_id         SMALLINT NOT NULL,
+    dish_id         INT NOT NULL,
     PRIMARY KEY (menu_id, dish_id),
+    
     FOREIGN KEY (menu_id) REFERENCES Menu(id),
     FOREIGN KEY (dish_id) REFERENCES Dishes(id)
 );
 GO
 
 CREATE TABLE Dish_Ingredients (
-    dish_id               SMALLINT NOT NULL,
-    ingredient_id           SMALLINT NOT NULL,
-    ingredient_quantity     SMALLINT,
-    unit                    VARCHAR(50),
+    dish_id               INT NOT NULL,
+    ingredient_id         INT NOT NULL,
+    ingredient_quantity   INT,
+    unit                  VARCHAR(50),
     PRIMARY KEY (dish_id, ingredient_id),
 
     FOREIGN KEY (dish_id) REFERENCES Dishes(id),
@@ -105,17 +106,18 @@ CREATE TABLE Dish_Ingredients (
 GO
 
 CREATE TABLE Orders (
-    id              INT PRIMARY KEY IDENTITY,
-    account_id      INT,
-    order_date      DATETIME,
-    ship_date       DATETIME,
-    ship_address    NVARCHAR(255),
-    ship_city_id    SMALLINT,
-    ship_district_id SMALLINT,
-    ship_ward_id    SMALLINT,
-    total_price     MONEY,
-    customer_notes  TEXT,
-    [status]        BIT NOT NULL DEFAULT 1,
+    id               INT PRIMARY KEY IDENTITY,
+    account_id       INT,
+    order_date       DATETIME,
+    ship_date        DATETIME,
+    ship_address     NVARCHAR(255),
+    ship_city_id     INT,
+    ship_district_id INT,
+    ship_ward_id     INT,
+    total_price      MONEY,
+    customer_notes   TEXT,
+    [status]         BIT NOT NULL DEFAULT 1,
+
     FOREIGN KEY (account_id) REFERENCES Accounts(id),
     FOREIGN KEY (ship_city_id) REFERENCES City(id),
     FOREIGN KEY (ship_district_id) REFERENCES District(id),
@@ -125,9 +127,9 @@ GO
 
 CREATE TABLE Order_Details (
     order_id        INT NOT NULL,
-    dish_id         SMALLINT NOT NULL,
+    dish_id         INT NOT NULL,
     item_type       BIT NOT NULL DEFAULT 1,
-    quantity        SMALLINT,
+    quantity        INT,
     price_per_unit  MONEY,
     PRIMARY KEY (order_id, dish_id),
     FOREIGN KEY (order_id) REFERENCES Orders(id),
